# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LuraDB API",
    "description": "REST-native multi-model database — KeyValue- und JSON-Engine. `version` ist die API-Contract-Version (siehe API_VERSION), unabhängig von der Server-Version; letztere steht in der Extension `x-luradb-server-version`. Kompatibilitäts-Ranges: api/COMPATIBILITY.md. Laufzeit-Check: GET /version.",
    "contact": {
      "name": "Heiko Wein"
    },
    "license": {
      "name": "FSL-1.1-ALv2"
    },
    "version": "0.1.0",
    "x-luradb-server-version": "0.1.0"
  },
  "paths": {
...
    "/store-api/domains": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Returns all active domains with their names and creation timestamps.",
        "operationId": "list_domains",
        "responses": {
          "200": {
            "description": "List of active domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DomainResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Creates a new logical domain (isolated KV namespace).",
        "description": "Domain names must be unique, max 50 chars, and match `[a-zA-Z0-9_-]`.",
        "operationId": "create_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain name"
          },
          "409": {
            "description": "Domain already exists"
          }
        }
      }
    },
    "/store-api/domains/{name}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Returns metadata for a single domain by name. Returns 404 if it does not exist.",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Schedules a domain and all its data for deletion.",
        "description": "The domain becomes immediately inaccessible; the actual purge runs in the background.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge started"
          },
          "404": {
            "description": "Domain not found"
          }
        }
      }
    },
    "/store-api/json/domains": {
      "get": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Returns all active JSON domains.",
        "operationId": "list_domains",
        "responses": {
          "200": {
            "description": "List of active JSON domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JsonDomainResponse"
                  }
                }
              }
            }
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      },
      "post": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Creates a new JSON domain (isolated document namespace, separate from KV domains).",
        "operationId": "create_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJsonDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonDomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain name"
          },
          "409": {
            "description": "Domain already exists"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      }
    },
    "/store-api/json/domains/{name}": {
      "get": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Returns metadata for a single JSON domain, including its document count.",
        "description": "Domains in `deleting` state answer with 410 Gone.",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "JSON domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonDomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is being deleted (background purge running)"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      },
      "delete": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Marks a JSON domain as deleting. Documents become inaccessible immediately.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "JSON domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge follows"
          },
          "404": {
            "description": "Domain not found"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      }
    },
...
    "/store-api/json/{domain}/documents/count": {
      "get": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Counts all documents of a domain (key scan only).",
        "operationId": "count_documents",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "JSON domain",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      }
    },
...
    "/store-api/rel/domains": {
      "get": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Returns all active relational domains.",
        "operationId": "list_domains",
        "responses": {
          "200": {
            "description": "List of active relational domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelDomainResponse"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      },
      "post": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Creates a new relational domain (isolated table/view namespace, separate",
        "description": "from KV and JSON domains).",
        "operationId": "create_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Relational domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelDomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain name"
          },
          "409": {
            "description": "Domain already exists"
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      }
    },
    "/store-api/rel/domains/{name}": {
      "get": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Returns metadata for a single relational domain. A domain in `deleting`",
        "description": "state answers 410 Gone (rel/013 finishes the physical purge).",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Relational domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelDomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is being deleted (background purge running)"
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      },
      "delete": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Marks a relational domain as deleting. Its tables/views become",
        "description": "inaccessible immediately; physical cleanup follows in rel/013.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Relational domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge follows (rel/013)"
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is already being deleted"
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountResponse": {
        "type": "object",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "CreateDomainRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "User-visible domain name (max 50 chars, [a-zA-Z0-9_-])."
          }
        }
      },
...
      "CreateJsonDomainRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "User-visible domain name (max 50 chars, [a-zA-Z0-9_-])."
          }
        }
      },
      "CreateRelDomainRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Domain name (max 50 chars, `[a-zA-Z0-9_-]`; `\"domains\"` is reserved)."
          }
        }
      },
...
      "DomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "User-visible domain name."
          }
        }
      },
...
      "JsonDomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at",
          "state"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "document_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of documents — only set on the detail endpoint.",
            "nullable": true,
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "Lifecycle state: \"active\" or \"deleting\" (background purge running)."
          }
        }
      },
...
      "RelDomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at",
          "state"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "Lifecycle state: \"active\" or \"deleting\" (background purge running, rel/013).\nAlways present — so rel/013's later `list_domains` visibility change\n(surfacing `Deleting` domains too) needs no REST-contract change."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key wie in `luradb.toml` konfiguriert (Admins) oder wie bei User-Anlage / Key-Rotation erhalten. Nur bei `auth.enabled = true` erforderlich."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}